Fix wasm
