+++ cpp-terminal/platforms/input.cpp
@@ -30,12 +30,64 @@
 #include "cpp-terminal/platforms/file.hpp"
 #include "cpp-terminal/platforms/input.hpp"
 
-#include <iostream>
 #include <string>
+#include <thread>
+
+namespace Term
+{
+namespace Private
+{
+template<typename T> class BlockingQueue
+{
+private:
+  std::mutex    mutex_;
+  std::queue<T> queue_;
+
+public:
+  T pop()
+  {
+    this->mutex_.lock();
+    T value = this->queue_.front();
+    this->queue_.pop();
+    this->mutex_.unlock();
+    return value;
+  }
+
+  void push(const T& value)
+  {
+    this->mutex_.lock();
+    this->queue_.push(value);
+    this->mutex_.unlock();
+  }
+
+  void push(T&& value)
+  {
+    this->mutex_.lock();
+    this->queue_.push(std::move(value));
+    this->mutex_.unlock();
+  }
+
+  bool empty()
+  {
+    this->mutex_.lock();
+    bool check = this->queue_.empty();
+    this->mutex_.unlock();
+    return check;
+  }
+  std::size_t size()
+  {
+    this->mutex_.lock();
+    std::size_t check = this->queue_.size();
+    this->mutex_.unlock();
+    return check;
+  }
+};
+}  // namespace Private
+}  // namespace Term
 
 std::thread Term::Private::Input::m_thread = std::thread(Term::Private::Input::read_event);
 
-Term::Private::Input::BlockingQueue<Term::Event> Term::Private::Input::m_events;
+Term::Private::BlockingQueue<Term::Event> Term::Private::Input::m_events;
 
 #if !defined(_WIN32) && !defined(__APPLE__)
 namespace Term
